Created PlayerStats script and related things
PlayerStats is basically the part of the player script the items would interact with. It's sort of a placeholder since a lot of the stats would probably be better served on a bigger player controller script, but for now this will work as something for the clothing to interact with.
Also I added support for non-clothing items (mostly because I was struggling with how the clothing should work in the inventory and what I have is part of the solution that I feel will work well)