Added debug script for PlayerStats and finished item equipment code
The debug script allows you to do various operations to the PlayerStats clothing, like adding or removing them. Everything on the PlayerStats side is finished, now I'll move on to the inventory.

Also I found a weird unity bug while testing: if I exit play mode while looking at a gameobject containing a list in the inspector, and that list's size changed during play mode (so it has to revert back to its original size when exiting play mode), then the list field breaks and disappears until I look at another object and look back, and any object with an integer field will flood the console with errors when I look at them in the inspector because something is trying to access the serialized object of the field after it's disposed. This fixes itself when reentering play mode or recompiling scripts.