Added inventory & inventory screen
It's set up so it should be relatively easy to add to the main scene. Only things we'd need is to add buttons for opening/closing the screen (the functions for doing those are already made) and implementing the DropItem() function in PlayerInventory, as well as adding both PlayerStats and PlayerInventory to the player character.